Revenue values being saved
Revenue values will now be calculated and saved after submitting import .xls. It's important to note that isolation is currently being ignored due to timerestraints. This means that the revenue will only include segment prices if the isDoneFloor collumn is set to true. Issue has been made to address this in the future.